move server type function to the batchq module
