+++ utilix/batchq.py
@@ -10,14 +10,29 @@ from pydantic import BaseModel, Field, validator
 from simple_slurm import Slurm  # type: ignore
 from . import logger
 
+# Get the SERVER type
+def get_server_type():
+    hostname = os.uname().nodename
+    if "midway2" in hostname:
+        return "Midway2"
+    elif "midway3" in hostname:
+        return "Midway3"
+    elif "dali" in hostname:
+        return "Dali"
+    else:
+        raise ValueError(
+            f"Unknown server type for hostname {hostname}. Please use midway2, midway3, or dali."
+        )
+        
+SERVER = get_server_type()
+
 USER: Optional[str] = os.environ.get("USER")
 if USER is None:
     raise ValueError("USER environment variable is not set")
 
 SCRATCH_DIR: str = os.environ.get("SCRATCH", ".")
-# SCRATCH_DIR must have write permission
-if not os.access(SCRATCH_DIR, os.W_OK):
-    if 'dali' not in os.environ.get("HOSTNAME"):
+# for non-dali hosts, SCRATCH_DIR must have write permission
+if not os.access(SCRATCH_DIR, os.W_OK) and SERVER != "Dali":
         raise ValueError(
             f"SCRATCH_DIR {SCRATCH_DIR} does not have write permission. "
             "You may need to set SCRATCH_DIR manually in your .bashrc or .bash_profile."
